Modify a Question Controller update and destroy methods for new Question Component by Vue.js

+++ app/Http/Controllers/QuestionsController.php
@@ -84,6 +84,14 @@ class QuestionsController extends Controller
     {
         $question->update($request->only('title', 'body'));
 
+        if ($request->expectsJson())
+        {
+            return response()->json([
+                'message' => "Your question has been updated.",
+                'body_html' => $question->body_html
+            ]);
+        }
+
         return redirect('/questions')->with('success', "Your question has been updated.");
     }
 
@@ -95,8 +103,17 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        $this->authorize("delete", $question);
+
         $question->delete();
 
+        if (request()->expectsJson())
+        {
+            return response()->json([
+                'message' => "Your question has been deleted."
+            ]);
+        }
 
+        return redirect('/question')->with('success', "Your question has been deleted.");
     }
 }
